Moved all interfaces to respective interface files
