+++ src/interfaces/TaskItem.ts
@@ -1,0 +1,9 @@
+interface TaskProps {
+	task: {
+		id: number;
+		body: string;
+		due: string;
+		tag: string;
+	};
+	onEdit: () => void;
+}
